backend/requirements.txt
django
gunicorn
djangorestframework
django-cors-headers
django-jsonfield
djangorestframework_simplejwt
pillow